[feat] 팀원들 코드 리뷰 반영 - 패키지명 log -> ddoeatlog 로 변경 - Follow 테이블에서 필드명 카멜케이스로 수정 - LikeIt 테이블명 -> Heart 로 수정
