+++ server-domain/src/main/java/com/depromeet/entity/ddoeatlog/DdoEatLog.java
@@ -1,4 +1,6 @@
-package com.depromeet.entity.like;
+package com.depromeet.entity.ddoeatlog;
+
+import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -12,15 +14,28 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class LikeIt {
+public class DdoEatLog {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long likeId;
+	private Long ddoeatlogId;
 
-	@Column(name = "image", columnDefinition = "varchar(500) null comment '이미지'")
+	@Column(nullable = false)
+	private Long storeId;
+
+	@Column(nullable = false)
 	private Long userId;
 
 	@Column(nullable = false)
-	private Long ddoeatlogId;
+	private Float rating;
+
+	@Column(nullable = false)
+	private LocalDateTime visitedAt;
+
+	private String imageUrl;
+
+	@Column(nullable = false)
+	private Integer visitTimes;
+
+	private String description;
 }
